move frontend to own plugin and create ui package, Create add new view dialog
