+++ packages/geoportia-ui/README.md
@@ -1,0 +1,12 @@
+# @internal/geoportia-ui
+
+_This package was created through the Backstage CLI_.
+
+## Installation
+
+Install the package via Yarn:
+
+```sh
+cd <package-dir> # if within a monorepo
+yarn add @internal/geoportia-ui
+```
