+++ plugins/catalog-frontend/src/components/Table/EntityBooleanValuePicker.tsx
