+++ plugins/catalog-frontend/src/index.ts
@@ -1,1 +1,4 @@
 export { catalogFrontendPlugin, CatalogFrontendPage } from './plugin';
+
+export { CustomTable } from './components/Table/CustomTable';
+export { EntityBooleanValuePicker } from './components/Table/EntityBooleanValuePicker';
